ConstructionNode.getOptionalModifiers() removed.
ontop-optimization is not yet compiling.

+++ core/model/src/main/java/it/unibz/inf/ontop/iq/node/LimitNode.java
@@ -2,7 +2,7 @@ package it.unibz.inf.ontop.iq.node;
 
 public interface LimitNode extends QueryModifierNode {
 
-    int getLimit();
+    long getLimit();
 
     @Override
     LimitNode clone();
