+++ core/model/src/main/java/it/unibz/inf/ontop/iq/node/OffsetNode.java
@@ -2,7 +2,7 @@ package it.unibz.inf.ontop.iq.node;
 
 public interface OffsetNode extends QueryModifierNode {
 
-    int getOffset();
+    long getOffset();
 
     @Override
     OffsetNode clone();
